requirements.txt
# ── LangChain ecosystem (versões pinadas e compatíveis entre si) ──────────
langchain==1.2.13
langchain-core==1.2.22
langchain-community==0.4.1
langchain-text-splitters==1.1.1
langchain-anthropic==1.4.0
langchain-classic==1.0.3
langchain-huggingface==1.2.1

# ── Embeddings locais (sem custo de API) ──────────────────────────────────
sentence-transformers==3.4.1

# ── Vector Store ──────────────────────────────────────────────────────────
chromadb==1.5.5

# ── Avaliação RAGAS ───────────────────────────────────────────────────────
ragas==0.4.3
datasets==4.8.4

# ── Dependências diretas do RAGAS (pinadas para evitar downgrade) ─────────
openai==2.29.0
langchain-openai==1.1.12

# ── Dados e utilitários ───────────────────────────────────────────────────
pandas==3.0.1
python-dotenv==1.2.2
pypdf==6.9.2
